feat: expose all tables via flight sql
This change moves the OLAP aggregation logic within ceramic-one. This
means that all tables are available via Flight SQL endpoint.
