+++ .gitignore
@@ -13,3 +13,4 @@ openapi-generator-cli.jar
 .env
 ceramic_cicddb.sqlite*
 event-svc/model_error_counts.csv
+/.history
