+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
     "kubo-rpc-server",
     "metadata",
     "metrics",
-    "olap",
     "one",
     "p2p",
     "pipeline",
@@ -148,6 +147,7 @@ multihash-derive = { version = "0.9" }
 names = { version = "0.14.0", default-features = false }
 nix = "0.26"
 num_enum = "0.5.7"
+object_store = { version = "0.11", features = ["aws"] }
 once_cell = "1.17.1"
 opentelemetry = "0.22"
 opentelemetry-otlp = "0.15"
