+++ event-svc/Cargo.toml
@@ -18,8 +18,8 @@ ceramic-core.workspace = true
 ceramic-event.workspace = true
 ceramic-metrics.workspace = true
 ceramic-sql.workspace = true
-ceramic-flight.workspace = true
 ceramic-validation.workspace = true
+ceramic-pipeline.workspace = true
 cid.workspace = true
 futures.workspace = true
 hex.workspace = true
