+++ event-svc/src/tests/event.rs
@@ -5,8 +5,7 @@ use anyhow::Error;
 use bytes::Bytes;
 use ceramic_api::{ApiItem, EventService as ApiEventService};
 use ceramic_core::NodeId;
-use ceramic_flight::server::ConclusionFeed as _;
-use ceramic_flight::ConclusionEvent;
+use ceramic_pipeline::{ConclusionEvent, ConclusionFeed as _};
 use ceramic_sql::sqlite::SqlitePool;
 use cid::{Cid, CidGeneric};
 use expect_test::expect;
