--- olap/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "ceramic-olap"
-description = "Standalone process that aggregates Ceramic events into a data lake for OLAP workloads"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-arrow.workspace = true
-ceramic-metrics.workspace = true
-ceramic-pipeline.workspace = true
-cid.workspace = true
-clap.workspace = true
-datafusion.workspace = true
-futures.workspace = true
-git-version = "0.3.9"
-hyper.workspace = true
-json-patch = "2.0.0"
-multibase.workspace = true
-multihash.workspace = true
-multihash-codetable = { version = "0.1.3", features = ["sha2"] }
-multihash-derive = "0.9.0"
-names.workspace = true
-object_store = { version = "0.11", features = ["aws"] }
-prometheus-client.workspace = true
-serde_json.workspace = true
-serde.workspace = true
-signal-hook = "0.3.17"
-signal-hook-tokio = { version = "0.3.1", features = ["futures-v0_3"] }
-tokio = { workspace = true, features = ["fs", "rt-multi-thread"] }
-tracing.workspace = true
-
-[dev-dependencies]
-test-log.workspace = true
-ceramic-flight.workspace = true
-expect-test.workspace = true
-ceramic-core.workspace = true
