+++ pipeline/Cargo.toml
@@ -10,11 +10,26 @@ repository.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-arrow-flight.workspace = true
+arrow-schema.workspace = true
+arrow.workspace = true
+async-stream.workspace = true
+async-trait.workspace = true
+ceramic-core.workspace = true
+ceramic-event.workspace = true
 cid.workspace = true
-datafusion-federation.workspace = true
-datafusion-flight-sql-table-provider.workspace = true
 datafusion.workspace = true
-object_store = { version = "0.11", features = ["aws"] }
-tonic.workspace = true
+expect-test.workspace = true
+futures.workspace = true
+int-enum.workspace = true
+ipld-core.workspace = true
+json-patch = "3.0.1"
+object_store.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
 url.workspace = true
+
+[dev-dependencies]
+test-log = "0.2.16"
+ceramic-arrow-test.workspace = true
