+++ pipeline/src/aggregator/ceramic_patch.rs
