+++ pipeline/src/conclusion/mod.rs
@@ -1,0 +1,8 @@
+mod event;
+mod table;
+
+pub use event::{
+    conclusion_events_to_record_batch, ConclusionData, ConclusionEvent, ConclusionInit,
+    ConclusionTime,
+};
+pub use table::{ConclusionFeed, FeedTable};
